requirements.txt
# Core HTTP client
requests>=2.31.0

# .env file support (load API credentials from a local file)
python-dotenv>=1.0.0

# urllib3 is a transitive dependency of requests — pinned for Retry stability
urllib3>=2.0.0
